# docs/swagger.yaml
definitions:
  apiAuth.ConfirmResetPasswordRequest:
    properties:
      new_password:
        example: newpassword123
        maxLength: 100
        minLength: 6
        type: string
      token:
        example: abc123...
        maxLength: 64
        minLength: 32
        type: string
    required:
    - new_password
    - token
    type: object
  apiAuth.LoginResponse:
    properties:
      created_at:
        type: string
      expired_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/entity.SystemUser'
    type: object
  apiAuth.LogoutResponse:
    properties:
      message:
        type: string
    type: object
  apiAuth.PasswordChangeRequest:
    properties:
      new_password:
        example: newpassword123
        maxLength: 100
        minLength: 6
        type: string
      old_password:
        example: oldpassword123
        maxLength: 100
        minLength: 6
        type: string
    required:
    - new_password
    - old_password
    type: object
  apiAuth.PasswordChangeResponse:
    properties:
      message:
        type: string
    type: object
  apiAuth.PasswordResetRequest:
    properties:
      email:
        example: admin@example.com
        type: string
    required:
    - email
    type: object
  apiAuth.PasswordResetResponse:
    properties:
      message:
        type: string
    type: object
  apiAuth.RegisterRequest:
    properties:
      email:
        example: admin@example.com
        type: string
      nickname:
        example: 筱锋
        maxLength: 50
        minLength: 1
        type: string
      password:
        example: password123
        maxLength: 100
        minLength: 6
        type: string
      username:
        example: admin
        maxLength: 50
        minLength: 1
        type: string
    required:
    - email
    - password
    - username
    type: object
  apiAuth.RegisterResponse:
    properties:
      created_at:
        type: string
      expired_at:
        type: string
      token:
        type: string
      user:
        $ref: '#/definitions/entity.SystemUser'
    type: object
  apiAuth.UserInfoResponse:
    properties:
      user:
        $ref: '#/definitions/entity.SystemUser'
    type: object
  apiInfo.AboutResponse:
    properties:
      content:
        type: string
      updated_at:
        type: string
    type: object
  apiInfo.AboutUpdateRequest:
    properties:
      content:
        example: |-
          # 关于我
          我是筱锋...
        maxLength: 10000
        minLength: 1
        type: string
    required:
    - content
    type: object
  apiInfo.SiteResponse:
    properties:
      introduction:
        type: string
      site_description:
        type: string
      site_name:
        type: string
      updated_at:
        type: string
    type: object
  apiInfo.SiteUpdateRequest:
    properties:
      introduction:
        example: 欢迎来到我的主页！
        maxLength: 2000
        type: string
      site_description:
        example: 一个有趣的个人博客
        maxLength: 500
        type: string
      site_name:
        example: 筱锋的小站
        maxLength: 100
        minLength: 1
        type: string
    type: object
...
  entity.SystemUser:
    properties:
      avatar:
        description: 头像URL
        type: string
      created_at:
        description: 创建时间
        type: string
      email:
        description: 邮箱
        type: string
      email_verify:
        description: 邮箱验证状态
        type: boolean
      id:
        description: ID 使用 Snowflake ID 作为主键
        type: integer
      last_login_at:
        description: 最后登录时间
        type: string
      nickname:
        description: 昵称
        type: string
      role:
        description: 角色
        type: string
      status:
        description: 状态
        type: integer
      updated_at:
        description: 更新时间
        type: string
      username:
        description: 用户名
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/v1/admin/info/about:
    put:
      consumes:
      - application/json
      description: 管理员更新 Markdown 格式的自我介绍
      parameters:
      - description: 自我介绍更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiInfo.AboutUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiInfo.AboutResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新自我介绍
      tags:
      - 站点信息管理
  /api/v1/admin/info/site:
    put:
      consumes:
      - application/json
      description: 管理员更新站点名称、描述、主页介绍
      parameters:
      - description: 站点信息更新请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiInfo.SiteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/apiInfo.SiteResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 更新站点信息
      tags:
      - 站点信息管理
...
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: 管理员用户登录，返回用户信息、访问令牌及Token时间信息
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，包含用户信息、Token及时间信息
          schema:
            $ref: '#/definitions/apiAuth.LoginResponse'
        "401":
          description: 认证失败
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 用户登录
      tags:
      - 认证管理
  /api/v1/auth/logout:
    patch:
      consumes:
      - application/json
      description: 注销当前登录会话
      produces:
      - application/json
      responses:
        "200":
          description: 登出成功
          schema:
            $ref: '#/definitions/apiAuth.LogoutResponse'
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 用户登出
      tags:
      - 认证管理
  /api/v1/auth/password/change:
    put:
      consumes:
      - application/json
      description: 修改当前用户的登录密码
      parameters:
      - description: 修改密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiAuth.PasswordChangeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/apiAuth.PasswordChangeResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未认证或旧密码错误
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 修改密码
      tags:
      - 认证管理
  /api/v1/auth/password/reset:
    patch:
      consumes:
      - application/json
      description: 通过邮箱重置用户密码，发送重置链接到邮箱
      parameters:
      - description: 重置密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiAuth.PasswordResetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 重置链接已发送
          schema:
            $ref: '#/definitions/apiAuth.PasswordResetResponse'
        "400":
          description: 请求参数错误
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 邮箱不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 重置密码
      tags:
      - 认证管理
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账户，注册成功后自动登录并返回访问令牌
      parameters:
      - description: 注册请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiAuth.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功，包含用户信息、Token及时间信息
          schema:
            $ref: '#/definitions/apiAuth.RegisterResponse'
        "400":
          description: 请求参数错误（用户名或邮箱已存在）
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - 认证管理
  /api/v1/auth/reset-password:
    get:
      consumes:
      - application/json
      description: 检查密码重置链接是否有效
      parameters:
      - description: 重置Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Token有效
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Token无效或已过期
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 验证重置密码Token
      tags:
      - 认证管理
    post:
      consumes:
      - application/json
      description: 通过重置Token设置新密码
      parameters:
      - description: 确认重置密码请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/apiAuth.ConfirmResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码重置成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Token无效或已过期
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 确认重置密码
      tags:
      - 认证管理
  /api/v1/auth/user:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的详细信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/apiAuth.UserInfoResponse'
        "401":
          description: 未认证
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: 获取用户信息
      tags:
      - 认证管理
  /api/v1/auth/verify-email:
    get:
      consumes:
      - application/json
      description: 通过邮箱中的验证链接验证用户邮箱
      parameters:
      - description: 验证Token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Token无效或已过期
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 验证邮箱
      tags:
      - 认证管理
  /api/v1/info/about:
    get:
      consumes:
      - application/json
      description: 获取 Markdown 格式的自我介绍内容
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiInfo.AboutResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取自我介绍
      tags:
      - 站点信息
  /api/v1/info/site:
    get:
      consumes:
      - application/json
      description: 获取站点名称、描述、主页介绍等公开信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            $ref: '#/definitions/apiInfo.SiteResponse'
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      summary: 获取站点信息
      tags:
      - 站点信息
...
swagger: "2.0"